+++ tests/testthat/test-ggcoefstats.R
@@ -7,9 +7,7 @@ test_that("ggcoefstats doesn't work if no estimate column found", {
 # default plots for each statistic ------------------------------------------
 
 test_that("default plots are rendered correctly for each type of statistic", {
-  skip_if_not_installed("survival")
-
-  library(survival)
+  skip_if_not_installed("withr")
 
   set.seed(123)
   expect_doppelganger(
@@ -29,14 +27,6 @@ test_that("default plots are rendered correctly for each type of statistic", {
     fig = ggcoefstats(stats::aov(wt ~ mpg * am, mtcars), effectsize.type = "omega")
   )
 
-  set.seed(123)
-  expect_doppelganger(
-    title = "chi2-statistic",
-    fig = suppressWarnings(ggcoefstats(
-      survival::coxph(Surv(time, status) ~ age + sex + frailty(inst), lung)
-    ))
-  )
-
   df <- as.data.frame(Titanic)
 
   mod_glm <- stats::glm(
@@ -50,6 +40,17 @@ test_that("default plots are rendered correctly for each type of statistic", {
   expect_doppelganger(
     title = "z-statistic",
     fig = ggcoefstats(mod_glm, conf.level = 0.90)
+  )
+
+  skip_if_not_installed("survival")
+  withr::local_package("survival")
+
+  set.seed(123)
+  expect_doppelganger(
+    title = "chi2-statistic",
+    fig = suppressWarnings(ggcoefstats(
+      survival::coxph(Surv(time, status) ~ age + sex + frailty(inst), lung)
+    ))
   )
 })
 
@@ -130,27 +131,7 @@ test_that(
 # edge cases -------------------------------------
 
 test_that(
-  desc = "missing values in numeric columns",
-  code = {
-    skip_if_not_installed("lme4")
-    skip_on_os(c("windows", "linux", "solaris"))
-
-    library(lme4, warn.conflicts = FALSE)
-    m_lmer <- ggcoefstats(lmer(Reaction ~ Days + (Days | Subject), data = sleepstudy))
-
-    expect_s3_class(m_lmer, "ggplot")
-
-    skip_on_os("mac", c("i386", "x86_64"))
-    set.seed(123)
-    expect_doppelganger(
-      title = "NAs in numeric columns",
-      fig = m_lmer
-    )
-  }
-)
-
-test_that(
-  desc = "edge cases",
+  desc = "works when CIs unavailable",
   code = {
     set.seed(123)
     df_base <- tidy_model_parameters(stats::lm(wt ~ am * cyl, mtcars))